api js -> ts
