+++ lib/api.ts
@@ -1,0 +1,44 @@
+import fs from "fs";
+import { join } from "path";
+import matter from "gray-matter";
+
+const postsDirectory = join(process.cwd(), "_posts");
+
+export function getPostSlugs(): string[] {
+  return fs.readdirSync(postsDirectory);
+}
+
+export function getPostBySlug(
+  slug: string,
+  fields: string[] = []
+): { [key: string]: any } {
+  const realSlug = slug.replace(/\.md$/, "");
+  const fullPath = join(postsDirectory, `${realSlug}.md`);
+  const fileContents = fs.readFileSync(fullPath, "utf8");
+  const { data, content } = matter(fileContents);
+
+  const items: { [key: string]: any } = {};
+
+  fields.forEach((field) => {
+    if (field === "slug") {
+      items[field] = realSlug;
+    }
+    if (field === "content") {
+      items[field] = content;
+    }
+
+    if (data[field]) {
+      items[field] = data[field];
+    }
+  });
+
+  return items;
+}
+
+export function getAllPosts(fields: string[] = []): any[] {
+  const slugs = getPostSlugs();
+  const posts = slugs
+    .map((slug) => getPostBySlug(slug, fields))
+    .sort((post1: any, post2: any) => (post1.date > post2.date ? -1 : 1));
+  return posts;
+}
